refactor: moves files and uses react hooks
separates logic to more easily observable chunks using hooks
