+++ src/index.css
@@ -1,6 +1,7 @@
 :root {
   --font-size: 1.3rem;
   --sans: sans-serif, 'Chivo';
+  --dark-blue: #1a2132;
 }
 
 *,
@@ -19,14 +20,16 @@ body {
 .container {
   display: flex;
   flex-direction: column;
-  justify-content: center;
+  color: white;
+  background-color: var(--dark-blue);
   align-items: center;
-  margin-top: 5rem;
+  padding-top: 5rem;
   gap: 1rem;
+  height: 100vh;
 }
 
 .timer-container {
-  border: 2px solid black;
+  border: 2px solid white;
   text-align: center;
   font-size: 4rem;
   padding: 2rem;
@@ -41,8 +44,8 @@ body {
   padding: 0.25rem 0.5rem;
   font-size: 1.5rem;
   margin: 0.25rem;
-  background-color: black;
-  color: white;
+  background-color: white;
+  color: darkblue;
   min-width: 5rem;
 }
 
@@ -52,6 +55,10 @@ body {
   text-align: center;
 }
 
+.lap-table th {
+  padding: 0.25rem;
+}
+
 .lap-table-header-row {
-  border-bottom: 1px solid black;
+  border-bottom: 1px solid white;
 }
