+++ src/stopwatch/StopWatch.tsx
@@ -1,10 +1,11 @@
-import React, { useCallback, useState } from 'react';
+import React, { useState } from 'react';
 import Timer from './Timer';
 import StopWatchButtons from './StopWatchButtons';
 import LapsList from './LapsList';
-import { StopWatchAction, StopWatchState } from './resources/stopWatch';
-import { toStopWatchFormat } from './utils/time';
+import { StopWatchState } from './resources/stopWatch';
+import { toStopWatchFormat } from '../utils/time';
 import { Lap } from './types';
+import { useHandleStopWatchAction } from './hooks/useHandleAction';
 
 export default function StopWatch() {
   const [laps, setLaps] = useState<Lap[]>([]);
@@ -12,32 +13,11 @@ export default function StopWatch() {
   const [lapNumber, setLapNumber] = useState<number>(0);
   const [state, setState] = useState<StopWatchState>(StopWatchState.INITIAL);
 
-  const handleAction = useCallback(
-    (event: StopWatchAction) => {
-      switch (event) {
-        case StopWatchAction.START:
-        case StopWatchAction.RESUME: {
-          setState(StopWatchState.RUNNING);
-          break;
-        }
-        case StopWatchAction.STOP: {
-          setState(StopWatchState.PAUSED);
-          break;
-        }
-        case StopWatchAction.LAP: {
-          setLapNumber((prevLapNumber) => prevLapNumber + 1);
-          break;
-        }
-        case StopWatchAction.RESET: {
-          setState(StopWatchState.INITIAL);
-          setLapNumber(0);
-          setPrevTime(0);
-          setLaps([]);
-          break;
-        }
-      }
-    },
-    [setState, setLapNumber, setPrevTime, setLaps]
+  const { handleAction } = useHandleStopWatchAction(
+    setState,
+    setLapNumber,
+    setPrevTime,
+    setLaps
   );
 
   const handleLap = (time: number) => {
