+++ src/stopwatch/StopWatchButtons.tsx
