+++ src/stopwatch/Timer.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { toStopWatchFormat } from './utils/time';
+import { toStopWatchFormat } from '../utils/time';
 import { StopWatchState } from './resources/stopWatch';
 
 interface IProps {
@@ -7,6 +7,9 @@ interface IProps {
   lapNumber: number;
   onLap: (time: number) => void;
 }
+
+// Attempts to minimize content rerendered when timer is updated
+// by rerendering only the timer portion
 
 export default function Timer({ state, lapNumber, onLap }: IProps) {
   const [time, setTime] = useState<number>(0);
@@ -16,7 +19,7 @@ export default function Timer({ state, lapNumber, onLap }: IProps) {
     if (state === StopWatchState.RUNNING) {
       intervalID = setInterval(() => {
         setTime((prevTime) => prevTime + 1);
-      }, 10);
+      }, 10); // increments timer every 10ms = 1/100 seconds
     }
     if (state === StopWatchState.INITIAL) {
       setTime(0);
@@ -24,6 +27,8 @@ export default function Timer({ state, lapNumber, onLap }: IProps) {
     if (intervalID) return () => clearInterval(intervalID);
   }, [state]);
 
+  // Triggers lap handler when lap button is pressed
+  // Does so by observing increases in lap #
   useEffect(() => {
     if (lapNumber !== 0) onLap(time);
   }, [lapNumber]);
