+++ src/stopwatch/resources/stopWatch.ts
