+++ src/stopwatch/types.ts
